refactor(notifee): move notifee to an exclusive file
