+++ src/app-content/index.js
@@ -1,9 +1,8 @@
 import {
-  useEffect, useRef, useState
+  useEffect, useRef
 } from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { NavigationContainer } from '@react-navigation/native';
-import notifee, { EventType } from '@notifee/react-native';
 import { useDispatch, useSelector } from 'react-redux';
 // providers and store
 import {
@@ -32,9 +31,7 @@ import {
   sessionStateChanged,
 } from '../store/redux/slices/wide-app/client';
 import logger from '../services/logger';
-import { toggleMuteMicrophone } from '../components/audio/service';
 import '../utils/locales/i18n';
-import Colors from '../constants/colors';
 
 //  Inject store in non-component files
 const injectStore = () => {
@@ -43,17 +40,6 @@ const injectStore = () => {
   injectStoreAM(store);
   injectStoreIM(store);
 };
-
-// Create the foreground service task runner/notification
-// this will keep mic working on Android when app is in background/sleep
-notifee.registerForegroundService((notification) => {
-  return new Promise(() => {
-    logger.debug({
-      logCode: 'app_service_notification',
-      extraInfo: { notification },
-    }, 'Foreground service notification started.');
-  });
-});
 
 const AppContent = ({
   onLeaveSession: _onLeaveSession,
@@ -64,8 +50,6 @@ const AppContent = ({
   const Stack = createNativeStackNavigator();
   const dispatch = useDispatch();
   const guestStatus = useSelector((state) => state.client.guestStatus);
-  const audioIsConnected = useSelector((state) => state.audio.isConnected);
-  const audioIsMuted = useSelector((state) => state.audio.isMuted);
   const leaveOnUnmount = useSelector((state) => {
     const { sessionState } = state.client;
     return !sessionState.ended
@@ -74,7 +58,6 @@ const AppContent = ({
           || sessionState.loggingIn
           || sessionState.loggingOut);
   });
-  const [notification, setNotification] = useState(null);
   const navigationRef = useRef(null);
   const leaveOnUnmountRef = useRef();
 
@@ -146,144 +129,14 @@ const AppContent = ({
   }, []);
 
   useEffect(() => {
-    if (audioIsConnected) {
-      // Start/show the notification foreground service
-      const getChannelIdAndDisplayNotification = async () => {
-        const channelId = await notifee.createChannel({
-          id: 'main_meeting_channel',
-          name: t('mobileSdk.notification.label'),
-          vibration: false,
-        });
-
-        const _notification = {
-          title: t('mobileSdk.notification.title'),
-          body: t('mobileSdk.notification.body'),
-          id: 'audio_foreground_notification',
-          android: {
-            channelId,
-            pressAction: {
-              id: 'default',
-            },
-            actions: [
-              {
-                title: t('app.leaveModal.confirm'),
-                pressAction: {
-                  id: 'leave',
-                },
-              },
-              audioIsMuted
-                ? {
-                  title: t('app.actionsBar.unmuteLabel'),
-                  pressAction: {
-                    id: 'unmute',
-                  },
-                } : {
-                  title: t('app.actionsBar.muteLabel'),
-                  pressAction: {
-                    id: 'mute',
-                  },
-                }
-            ],
-            asForegroundService: true,
-            ongoing: true,
-            color: Colors.blue,
-            colorized: true,
-            smallIcon: 'ic_launcher_foreground',
-          },
-          ios: {
-            foregroundPresentationOptions: {
-              banner: false,
-            },
-          }
-        };
-        notifee.displayNotification(_notification);
-        setNotification(_notification);
-      };
-      getChannelIdAndDisplayNotification();
-    } else {
-      // stop notification service
-      notifee.stopForegroundService();
-    }
-  }, [audioIsConnected]);
-
-  useEffect(() => {
-    const requestNotificationPermission = async () => {
-      await notifee.requestPermission();
-    };
-
-    requestNotificationPermission();
-  }, []);
-
-  useEffect(() => {
-    if (notification) {
-      if (audioIsMuted) {
-        notification.android.actions[1].pressAction.id = 'unmute';
-        notification.android.actions[1].title = t('app.actionsBar.unmuteLabel');
-        notifee.displayNotification(notification);
-        setNotification(notification);
-      } else {
-        notification.android.actions[1].pressAction.id = 'mute';
-        notification.android.actions[1].title = t('app.actionsBar.muteLabel');
-        notifee.displayNotification(notification);
-        setNotification(notification);
-      }
-    }
-  }, [audioIsMuted]);
-
-  useEffect(() => {
-    // Background event = device locked || app not in view || killed/quit
-    notifee.onBackgroundEvent(async (event) => {
-      const { type, detail } = event;
-
-      if (detail.notification?.android?.channelId !== 'main_meeting_channel') {
-        return;
-      }
-
-      if (type === EventType.ACTION_PRESS && detail.pressAction.id === 'leave') {
-        dispatch(leave(api));
-
-        // Remove the notification
-        await notifee.cancelNotification(detail.notification.id);
-        return;
-      }
-
-      // Check if the user pressed the Mute/Unmute button
-      if (type === EventType.ACTION_PRESS && (detail.pressAction.id === 'mute' || detail.pressAction.id === 'unmute')) {
-        toggleMuteMicrophone();
-      }
-    });
-    return () => {
-      notifee.stopForegroundService();
-    };
-  }, []);
-
-  useEffect(() => {
     // Inject custom provided onLeaveSession callback into the store so it's called
     // once the store's about to be flushed
     if (typeof _onLeaveSession === 'function') injectStoreFlushCallback(_onLeaveSession);
     injectStore();
     dispatch(ConnectionStatusTracker.registerConnectionStatusListeners());
-    // Foreground event = device unlocked || app in view
-    const unsubscribeForegroundEvents = notifee.onForegroundEvent((event) => {
-      const { type, detail } = event;
-
-      if (detail.notification?.android?.channelId !== 'main_meeting_channel') {
-        return;
-      }
-
-      if (type === EventType.ACTION_PRESS && detail.pressAction.id === 'leave') {
-        dispatch(leave(api));
-        return;
-      }
-      if (type === EventType.ACTION_PRESS && (detail.pressAction.id === 'mute' || detail.pressAction.id === 'unmute')) {
-        toggleMuteMicrophone();
-      }
-    });
 
     return () => {
       dispatch(ConnectionStatusTracker.unregisterConnectionStatusListeners());
-      unsubscribeForegroundEvents();
-
       if (leaveOnUnmountRef.current) {
         dispatch(leave(api)).unwrap().finally(() => {
           dispatch(sessionStateChanged({
